Correct VAT and postal code for Campione d'Italia (#168)
* Campione d'Italia, IT became part of the EU customs territory on 1 Jan 2020

Campione d’Italia formally became part of the EU customs territory on 1 January 2020. VAT will apply, but the tax rate will remain that of Switzerland. Also, the Swiss postal code ceased to be valid.

* Revert the Campione d'Italia VAT rate exception to zero

It seems that the tax is not VAT as such, but rather "a local consumption tax [...] which is in line with Swiss VAT".

https://www.admin.ch/gov/en/start/documentation/media-releases.msg-id-77668.html

"territories excluded from the application of the VAT."
https://www.agenziaentrate.gov.it/portale/web/english/nse/services/vat-mini-one-stop-shop/faq/general-topics